refactor: type and ux improvements on translations
