--- apps/client/src/translation/Translation.types.ts
@@ -1,1 +1,0 @@
-export type TranslationObject = { [key: string]: string };
